requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.25.2
yfinance==0.2.28
alpha-vantage==2.3.1
polygon-api-client==1.14.0
websocket-client==1.6.4
aiohttp==3.9.1
asyncio==3.4.3
schedule==1.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.16.1
scikit-learn==1.3.2
ta-lib==0.4.28
quantlib==1.32
ccxt==4.1.68
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
lxml==4.9.3
selenium==4.15.2
feedparser==6.0.10
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8
tweepy==4.14.0
python-telegram-bot==20.7
discord.py==2.3.2